requirements.txt
pandas
tiktoken
numpy
openai
tenacity
streamlit
fpdf2